% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{new_basic_gt_block}
\alias{new_basic_gt_block}
\title{Create a new basic GT block}
\usage{
new_basic_gt_block(
  title = character(),
  subtitle = character(),
  footnotes = character(),
  ...
)
}
\arguments{
\item{title, subtitle, footnotes}{Initial text for each field (character).
All default to empty and accept markdown formatting.}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\value{
A basic gt block object that can be used with the serve function.
}
\description{
Creates a new GT block with inputs for title, subtitle, and footnotes. This
block provides an interface for a basic GT table.
}
\examples{
\dontrun{
serve(new_basic_gt_block(title = "test"), data = list(data = mtcars))
}

}
