% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{new_colour_gt_block}
\alias{new_colour_gt_block}
\title{Color a gt block}
\usage{
new_colour_gt_block(
  columns = numeric(),
  rows = character(),
  direction = character(),
  method = character(),
  palette = character(),
  bins = integer(),
  quantiles = integer(),
  alpha = numeric(),
  reverse = logical(),
  apply_to = character(),
  ...
)
}
\arguments{
\item{columns}{The columns to which cell data color operations are
constrained.}

\item{rows}{In conjunction with columns, we can specify which rows should
form a constraint for cell data color operations. Defaults to everything.}

\item{direction}{Should the color computations be performed column-wise
(default) or row-wise? Defeaults to colour down columns.}

\item{method}{A method for computing color based on the data within body
cells. Can be "auto" (the default), "numeric", "bin", "quantile", or
"factor". The "auto" method will automatically choose the "numeric" method
for numerical input data or the "factor" method for any non-numeric inputs.}

\item{palette}{Name of the colour palette to use.}

\item{bins}{Number of bins to use when method is "bin".}

\item{quantiles}{Number of equal-sized quantiles to use when method is
"quantile".}

\item{alpha}{Transparency level for colors.}

\item{reverse}{Logical indicating whether colors should be computed in
reverse order.}

\item{apply_to}{Whether colors should fill the cell background or the text.}

\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}
}
\value{
A gt block with additional cell colouring.
}
\description{
Color targeted cells in a GT block.
}
\examples{
\dontrun{
serve(new_colour_gt_block(), data = list(gt_obj = gt(head(mtcars))))
}

}
