% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{new_complete_gt_block}
\alias{new_complete_gt_block}
\title{Create a new complete GT block}
\usage{
new_complete_gt_block(
  title = character(),
  subtitle = character(),
  footnotes = character()
)
}
\arguments{
\item{title, subtitle, footnotes}{Initial text for each field (character).
All default to empty and accept markdown formatting.}
}
\value{
A transform block object that can be used with the serve function.
}
\description{
Creates a new GT block with inputs for title, subtitle, and footnotes. This
block provides a complete interface for a basic GT table.
}
\examples{
\dontrun{
serve(new_complete_gt_block(), list(data = mtcars))
}

}
