% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner.R
\name{new_spanner_gt_block}
\alias{new_spanner_gt_block}
\title{Add a spanner to a gt table}
\usage{
new_spanner_gt_block(label = character(), columns = character(), ...)
}
\arguments{
\item{label}{The text to use for the spanner label. Accepts markdown
formatted text.}

\item{columns}{The columns to serve as components of the spanner.}

\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}
}
\value{
A gt block object with an added spanner.
}
\description{
Add a spanner to a gt table
}
\examples{
\dontrun{
serve(new_spanner_gt_block(), data = list(gt_obj = gt(mtcars)))
}

}
