% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{new_style_gt_block}
\alias{new_style_gt_block}
\title{Style a GT block}
\usage{
new_style_gt_block(style = 1, color = "gray", striping = "yes", ...)
}
\arguments{
\item{style}{Numeric from 1 to 6 indicating which prebuilt default to use.}

\item{color}{One of "blue", "cyan", "pink", "green", "red", and "gray".}

\item{striping}{Logical indicating whether optional row striping
should be enabled.}

\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}
}
\value{
A styled gt block object that can be used with the serve function.
}
\description{
Style a GT block using one of six prebuilt defaults with further options to
control the table colour and row striping.
}
\examples{
\dontrun{
serve(new_style_gt_block(), data = list(gt_obj = gt(head(mtcars))))
}

}
