% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{new_style_gt_block}
\alias{new_style_gt_block}
\title{Style a GT block}
\usage{
new_style_gt_block(
  style = numeric(),
  color = character(),
  add_row_striping = logical()
)
}
\arguments{
\item{style}{Numeric from 1 to 6 indicating which prebuilt default to use.}

\item{color}{One of "blue", "cyan", "pink", "green", "red", and "gray".}

\item{add_row_striping}{Logical indicating whether optional row striping
should be enabled.}
}
\value{
A styled gt block object that can be used with the serve function.
}
\description{
Style a GT block using one of six prebuilt defaults with further options to
control the table colour and row striping.
}
\examples{
\dontrun{
serve(new_style_gt_block(), list(gt_obj = gt::gt(head(mtcars))))
}

}
